# pkg/api/oas_schemas_gen.go
// Code generated by ogen, DO NOT EDIT.

package api

// Ref: #/components/schemas/CreateSessionRequestSchema
type CreateSessionRequestSchema struct {
	// Encoded public key.
	EncodedPublicKey []byte `json:"encoded_public_key"`
}

// GetEncodedPublicKey returns the value of EncodedPublicKey.
func (s *CreateSessionRequestSchema) GetEncodedPublicKey() []byte {
	return s.EncodedPublicKey
}

// SetEncodedPublicKey sets the value of EncodedPublicKey.
func (s *CreateSessionRequestSchema) SetEncodedPublicKey(val []byte) {
	s.EncodedPublicKey = val
}

// Ref: #/components/schemas/CreateSessionResponseSchema
type CreateSessionResponseSchema struct {
	// Session.
	Session string `json:"session"`
	// Encoded public key.
	EncodedPublicKey []byte `json:"encoded_public_key"`
}

// GetSession returns the value of Session.
func (s *CreateSessionResponseSchema) GetSession() string {
	return s.Session
}

// GetEncodedPublicKey returns the value of EncodedPublicKey.
func (s *CreateSessionResponseSchema) GetEncodedPublicKey() []byte {
	return s.EncodedPublicKey
}

// SetSession sets the value of Session.
func (s *CreateSessionResponseSchema) SetSession(val string) {
	s.Session = val
}

// SetEncodedPublicKey sets the value of EncodedPublicKey.
func (s *CreateSessionResponseSchema) SetEncodedPublicKey(val []byte) {
	s.EncodedPublicKey = val
}

func (*CreateSessionResponseSchema) createSessionRes() {}

// Ref: #/components/schemas/ErrorResponseSchema
type ErrorResponseSchema struct {
	Message string `json:"message"`
}

// GetMessage returns the value of Message.
func (s *ErrorResponseSchema) GetMessage() string {
	return s.Message
}

// SetMessage sets the value of Message.
func (s *ErrorResponseSchema) SetMessage(val string) {
	s.Message = val
}

func (*ErrorResponseSchema) createSessionRes()  {}
func (*ErrorResponseSchema) submitPasswordRes() {}

// SubmitPasswordCreated is response for SubmitPassword operation.
type SubmitPasswordCreated struct{}

func (*SubmitPasswordCreated) submitPasswordRes() {}

// Ref: #/components/schemas/SubmitPasswordRequestSchema
type SubmitPasswordRequestSchema struct {
	// Session.
	Session string `json:"session"`
	// Encrypted password.
	EncryptedPassword string `json:"encrypted_password"`
}

// GetSession returns the value of Session.
func (s *SubmitPasswordRequestSchema) GetSession() string {
	return s.Session
}

// GetEncryptedPassword returns the value of EncryptedPassword.
func (s *SubmitPasswordRequestSchema) GetEncryptedPassword() string {
	return s.EncryptedPassword
}

// SetSession sets the value of Session.
func (s *SubmitPasswordRequestSchema) SetSession(val string) {
	s.Session = val
}

// SetEncryptedPassword sets the value of EncryptedPassword.
func (s *SubmitPasswordRequestSchema) SetEncryptedPassword(val string) {
	s.EncryptedPassword = val
}